+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/boundary/SessionClient.kt
@@ -1,0 +1,8 @@
+package com.nabla.sdk.core.domain.boundary
+
+public interface SessionClient {
+    public fun initSession(sessionTokenProvider: SessionTokenProvider)
+    public fun isSessionInitialized(): Boolean
+    public suspend fun getFreshAccessToken(forceRefreshAccessToken: Boolean = false): String
+    public fun clearSession()
+}