+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/boundary/TokenRepository.kt
@@ -1,7 +1,0 @@
-package com.nabla.sdk.core.domain.boundary
-
-internal interface TokenRepository {
-    fun initSession(sessionTokenProvider: SessionTokenProvider)
-    suspend fun getFreshAccessToken(forceRefreshAccessToken: Boolean = false): String
-    fun clearSession()
-}