+++ messaging-core/src/main/java/com/nabla/sdk/core/domain/entity/NablaException.kt
@@ -7,8 +7,6 @@ public sealed class NablaException private constructor(
     message: String? = null,
     cause: Throwable? = null,
 ) : Exception(message, cause) {
-    internal open val wrapped: Throwable? = null
-
     public sealed class Configuration(message: String) : NablaException(message = message) {
         public object MissingInitialize : Configuration("Missing SDK initialize. Make sure you call \"NablaCore.initialize\".")
         public object MissingApiKey :
@@ -17,13 +15,16 @@ public sealed class NablaException private constructor(
         public object MissingContext : Configuration("Missing context. Make sure you follow the doc to integrate the SDK properly.")
     }
 
-    public class Authentication internal constructor(override val wrapped: Throwable) : NablaException()
+    public sealed class Authentication constructor(cause: Throwable?, message: String) : NablaException(cause = cause, message = message) {
+        public object NotAuthenticated : Authentication(cause = null, message = "You must call NablaClient.authenticate before using any authenticated API")
+        public class UnableToGetFreshSessionToken(cause: Throwable) : Authentication(cause = cause, message = "Unable to get session token from the SessionTokenProvider")
+    }
 
-    public class Network internal constructor(override val wrapped: Throwable) : NablaException()
-    public class Server internal constructor(override val wrapped: Throwable, code: Int, serverMessage: String, requestId: String?) :
-        NablaException(message = "Nabla server error. Code: $code, message: $serverMessage, requestId: $requestId")
+    public class Network internal constructor(cause: Throwable) : NablaException(cause = cause)
+    public class Server internal constructor(cause: Throwable, code: Int, serverMessage: String, requestId: String?) :
+        NablaException(cause = cause, message = "Nabla server error. Code: $code, message: $serverMessage, requestId: $requestId")
 
-    public class Internal constructor(override val wrapped: Throwable) : NablaException(message = wrapped.message)
+    public class Internal constructor(cause: Throwable) : NablaException(cause = cause, message = cause.message)
 
     public class InvalidMessage(message: String) : NablaException(message = message)
     public class MessageNotFound(conversationId: ConversationId, localMessageId: MessageId.Local) : NablaException(
@@ -36,5 +37,5 @@ public sealed class NablaException private constructor(
     public class InvalidAppTheme(message: String) :
         NablaException(message = "$message. Please make sure you're using \"Theme.Material3\" on your app theme or at least for activities containing Nabla UI components.")
 
-    public class Unknown internal constructor(override val wrapped: Throwable) : NablaException(message = wrapped.message, cause = wrapped)
+    public class Unknown internal constructor(cause: Throwable) : NablaException(message = cause.message, cause = cause)
 }