+++ messaging-core/src/main/java/com/nabla/sdk/messaging/core/injection/MessagingContainer.kt
@@ -4,6 +4,7 @@ import com.apollographql.apollo3.ApolloClient
 import com.nabla.sdk.core.data.exception.NablaExceptionMapper
 import com.nabla.sdk.core.domain.boundary.FileUploadRepository
 import com.nabla.sdk.core.domain.boundary.Logger
+import com.nabla.sdk.core.domain.boundary.SessionClient
 import com.nabla.sdk.messaging.core.data.apollo.GqlMapper
 import com.nabla.sdk.messaging.core.data.conversation.ConversationRepositoryImpl
 import com.nabla.sdk.messaging.core.data.conversation.GqlConversationDataSource
@@ -21,6 +22,7 @@ internal class MessagingContainer(
     apolloClient: ApolloClient,
     fileUploadRepository: FileUploadRepository,
     val nablaExceptionMapper: NablaExceptionMapper,
+    val sessionClient: SessionClient,
 ) {
     private val repoScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
     private val gqlMapper = GqlMapper(logger)